+++ extensions/readwise-to-tana/src/book.tsx
@@ -55,13 +55,11 @@ export default function Book({ id, template }: BookProps) {
       ...highlight,
       updated: formatDate(highlight.updated),
       highlighted_at: formatDate(highlight.highlighted_at),
-      note: (highlight.note ?? '')
-        .split('\n')
-        .filter((note) => note)
-        .map((note) => `\n      - ${note}`)
-        .join('\n'),
+      note: (highlight.note ?? '').split('\n').filter((note) => note),
     })),
   })
+
+  console.log(allHighlights)
 
   const handleCopyAll = async () => {
     const currentTime = new Date().toISOString()